+++ R/TADPole.R
@@ -54,7 +54,7 @@ TADPole <- function(data, window.size = NULL, k = 2, dc, error.check = TRUE) {
      if (n < 2)
           stop("data should have more than one time series")
      if (k >= n)
-          stop("Number of clusters should be less that the number of time series")
+          stop("Number of clusters should be less than the number of time series")
 
      ## Calculate matrices with bounds
      LBM <- proxy::dist(x, x, method = "LBK",
@@ -110,6 +110,7 @@ TADPole <- function(data, window.size = NULL, k = 2, dc, error.check = TRUE) {
 
      if (length(ind1) > 0) {
           d1 <- proxy::dist(x[ind1[, 1]], x[ind1[, 2]], method = "DTW2",
+                            step.pattern = symmetric1,
                             window.type = "slantedband", window.size = window.size,
                             pairwise = TRUE)
 
@@ -187,6 +188,7 @@ TADPole <- function(data, window.size = NULL, k = 2, dc, error.check = TRUE) {
 
           if (sum(indCompute) > 0) {
                d2 <- proxy::dist(x[ii], x[indHDN[indCompute]], method = "DTW2",
+                                 step.pattern = symmetric1,
                                  window.type = "slantedband", window.size = window.size)
 
                delta[indCompute] <- d2